Drop label restriction on matrix lightweight executor
The parent matrix job does not really perform any function other than to
shepherd the children job matrices, which is where the work is done.
This lightweight parent job can run anywhere. The label restriction was
taking up a non-executing slot on the slave and preventing any other
concurrency throttled jobs from running on the slave.