Jenkins jobs are to clean up tmp files and the workspace folder when they are finished.
Jobs were generating a lot of test data under /tmp that wasn't getting cleaned up, as well as 1GB+ workspaces because of build directories.
And because the jobs in the pipeline don't re-use the built artifact for that SHA (from the first pipeline step), each job clones and builds its own cassandra codebase.

ref: https://issues.apache.org/jira/browse/INFRA-19793