Add jenkins->jira integration script
This script will parse a jenkins job of test failures and format a comment to
be added to JIRA about the build run, test failures, and attempt to link to
existing JIRA tickets around those failures. Functionality exists in the script
(as yet untested) to automate updating tickets with the results of this
parsing; we'll need to determine where this is going to run in what contexts as
well as set up an account for it to post from (JenkinsBot?) and then test and
commit that change in tandem.